Refactor: move special tx payload validation to specialtx
thus making providertx responsible only of the definition of the payload
primitives (therefore not depending on deterministicmns).

this breaks several circular dependencies.
