+++ src/evo/providertx.h
@@ -7,15 +7,9 @@
 
 #include "bls/bls_wrapper.h"
 #include "primitives/transaction.h"
-#include "consensus/validation.h"
-
 #include "netaddress.h"
-#include "pubkey.h"
 
 #include <univalue.h>
-
-class CBlockIndex;
-class CCoinsViewCache;
 
 // Provider-Register tx payload
 
@@ -163,13 +157,8 @@ public:
     void ToJson(UniValue& obj) const;
 };
 
-bool CheckProRegTx(const CTransaction& tx, const CBlockIndex* pindexPrev, const CCoinsViewCache* view, CValidationState& state);
-bool CheckProUpServTx(const CTransaction& tx, const CBlockIndex* pindexPrev, CValidationState& state);
-bool CheckProUpRegTx(const CTransaction& tx, const CBlockIndex* pindexPrev, const CCoinsViewCache* view, CValidationState& state);
-bool CheckProUpRevTx(const CTransaction& tx, const CBlockIndex* pindexPrev, CValidationState& state);
 
 // If tx is a ProRegTx, return the collateral outpoint in outRet.
 bool GetProRegCollateral(const CTransactionRef& tx, COutPoint& outRet);
-
 
 #endif  //PIVX_PROVIDERTX_H
