+++ src/evo/specialtx.cpp
@@ -6,13 +6,416 @@
 #include "evo/specialtx.h"
 
 #include "chain.h"
+#include "coins.h"
 #include "chainparams.h"
 #include "clientversion.h"
 #include "consensus/validation.h"
 #include "evo/deterministicmns.h"
 #include "evo/providertx.h"
+#include "messagesigner.h"
 #include "primitives/transaction.h"
 #include "primitives/block.h"
+#include "script/standard.h"
+
+/* -- Helper static functions -- */
+
+static bool CheckService(const CService& addr, CValidationState& state)
+{
+    if (!addr.IsValid()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-ipaddr");
+    }
+    if (!Params().IsRegTestNet() && !addr.IsRoutable()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-ipaddr");
+    }
+
+    // IP port must be the default one on main-net, which cannot be used on other nets.
+    static int mainnetDefaultPort = CreateChainParams(CBaseChainParams::MAIN)->GetDefaultPort();
+    if (Params().NetworkIDString() == CBaseChainParams::MAIN) {
+        if (addr.GetPort() != mainnetDefaultPort) {
+            return state.DoS(10, false, REJECT_INVALID, "bad-protx-ipaddr-port");
+        }
+    } else if (addr.GetPort() == mainnetDefaultPort) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-ipaddr-port");
+    }
+
+    // !TODO: add support for IPv6 and Tor
+    if (!addr.IsIPv4()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-ipaddr");
+    }
+
+    return true;
+}
+
+template <typename Payload>
+static bool CheckHashSig(const Payload& pl, const CKeyID& keyID, CValidationState& state)
+{
+    std::string strError;
+    if (!CHashSigner::VerifyHash(::SerializeHash(pl), keyID, pl.vchSig, strError)) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-sig", false, strError);
+    }
+    return true;
+}
+
+template <typename Payload>
+static bool CheckHashSig(const Payload& pl, const CBLSPublicKey& pubKey, CValidationState& state)
+{
+    if (!pl.sig.VerifyInsecure(pubKey, ::SerializeHash(pl))) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-sig", false);
+    }
+    return true;
+}
+
+template <typename Payload>
+static bool CheckStringSig(const Payload& pl, const CKeyID& keyID, CValidationState& state)
+{
+    std::string strError;
+    if (!CMessageSigner::VerifyMessage(keyID, pl.vchSig, pl.MakeSignString(), strError)) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-sig", false, strError);
+    }
+    return true;
+}
+
+template <typename Payload>
+static bool CheckInputsHash(const CTransaction& tx, const Payload& pl, CValidationState& state)
+{
+    if (CalcTxInputsHash(tx) != pl.inputsHash) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-inputs-hash");
+    }
+
+    return true;
+}
+
+static bool CheckCollateralOut(const CTxOut& out, const ProRegPL& pl, CValidationState& state, CTxDestination& collateralDestRet)
+{
+    if (!ExtractDestination(out.scriptPubKey, collateralDestRet)) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-collateral-dest");
+    }
+    // don't allow reuse of collateral key for other keys (don't allow people to put the collateral key onto an online server)
+    // this check applies to internal and external collateral, but internal collaterals are not necessarely a P2PKH
+    if (collateralDestRet == CTxDestination(pl.keyIDOwner) ||
+            collateralDestRet == CTxDestination(pl.keyIDVoting)) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-collateral-reuse");
+    }
+    // check collateral amount
+    if (out.nValue != Params().GetConsensus().nMNCollateralAmt) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-collateral-amount");
+    }
+    return true;
+}
+
+// Provider Register Payload
+static bool CheckProRegTx(const CTransaction& tx, const CBlockIndex* pindexPrev, const CCoinsViewCache* view, CValidationState& state)
+{
+    assert(tx.nType == CTransaction::TxType::PROREG);
+
+    ProRegPL pl;
+    if (!GetTxPayload(tx, pl)) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-payload");
+    }
+
+    if (pl.nVersion == 0 || pl.nVersion > ProRegPL::CURRENT_VERSION) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-version");
+    }
+    if (pl.nType != 0) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-type");
+    }
+    if (pl.nMode != 0) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-mode");
+    }
+
+    if (pl.keyIDOwner.IsNull() || pl.keyIDVoting.IsNull()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-key-null");
+    }
+    if (!pl.pubKeyOperator.IsValid()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-operator-key-invalid");
+    }
+    // we may support other kinds of scripts later, but restrict it for now
+    if (!pl.scriptPayout.IsPayToPublicKeyHash()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-payee");
+    }
+    if (!pl.scriptOperatorPayout.empty() && !pl.scriptOperatorPayout.IsPayToPublicKeyHash()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-operator-payee");
+    }
+
+    CTxDestination payoutDest;
+    if (!ExtractDestination(pl.scriptPayout, payoutDest)) {
+        // should not happen as we checked script types before
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-payee-dest");
+    }
+    // don't allow reuse of payout key for other keys (don't allow people to put the payee key onto an online server)
+    if (payoutDest == CTxDestination(pl.keyIDOwner) ||
+            payoutDest == CTxDestination(pl.keyIDVoting)) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-payee-reuse");
+    }
+
+    // It's allowed to set addr to 0, which will put the MN into PoSe-banned state and require a ProUpServTx to be issues later
+    // If any of both is set, it must be valid however
+    if (pl.addr != CService() && !CheckService(pl.addr, state)) {
+        // pass the state returned by the function above
+        return false;
+    }
+
+    if (pl.nOperatorReward > 10000) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-operator-reward");
+    }
+
+    if (pl.collateralOutpoint.hash.IsNull()) {
+        // collateral included in the proReg tx
+        if (pl.collateralOutpoint.n >= tx.vout.size()) {
+            return state.DoS(10, false, REJECT_INVALID, "bad-protx-collateral-index");
+        }
+        CTxDestination collateralTxDest;
+        if (!CheckCollateralOut(tx.vout[pl.collateralOutpoint.n], pl, state, collateralTxDest)) {
+            // pass the state returned by the function above
+            return false;
+        }
+        // collateral is part of this ProRegTx, so we know the collateral is owned by the issuer
+        if (!pl.vchSig.empty()) {
+            return state.DoS(100, false, REJECT_INVALID, "bad-protx-sig");
+        }
+    } else if (pindexPrev != nullptr) {
+        assert(view != nullptr);
+
+        // Referenced external collateral.
+        // This is checked only when pindexPrev is not null (thus during ConnectBlock-->CheckSpecialTx),
+        // because this is a contextual check: we need the updated utxo set, to verify that
+        // the coin exists and it is unspent.
+        Coin coin;
+        if (!view->GetUTXOCoin(pl.collateralOutpoint, coin)) {
+            return state.DoS(10, false, REJECT_INVALID, "bad-protx-collateral");
+        }
+        CTxDestination collateralTxDest;
+        if (!CheckCollateralOut(coin.out, pl, state, collateralTxDest)) {
+            // pass the state returned by the function above
+            return false;
+        }
+        // Extract key from collateral. This only works for P2PK and P2PKH collaterals and will fail for P2SH.
+        // Issuer of this ProRegTx must prove ownership with this key by signing the ProRegTx
+        const CKeyID* keyForPayloadSig = boost::get<CKeyID>(&collateralTxDest);
+        if (!keyForPayloadSig) {
+            return state.DoS(10, false, REJECT_INVALID, "bad-protx-collateral-pkh");
+        }
+        // collateral is not part of this ProRegTx, so we must verify ownership of the collateral
+        if (!CheckStringSig(pl, *keyForPayloadSig, state)) {
+            // pass the state returned by the function above
+            return false;
+        }
+    }
+
+    if (!CheckInputsHash(tx, pl, state)) {
+        return false;
+    }
+
+    if (pindexPrev) {
+        auto mnList = deterministicMNManager->GetListForBlock(pindexPrev);
+        // only allow reusing of addresses when it's for the same collateral (which replaces the old MN)
+        if (mnList.HasUniqueProperty(pl.addr) && mnList.GetUniquePropertyMN(pl.addr)->collateralOutpoint != pl.collateralOutpoint) {
+            return state.DoS(10, false, REJECT_DUPLICATE, "bad-protx-dup-IP-address");
+        }
+        // never allow duplicate keys, even if this ProTx would replace an existing MN
+        if (mnList.HasUniqueProperty(pl.keyIDOwner)) {
+            return state.DoS(10, false, REJECT_DUPLICATE, "bad-protx-dup-owner-key");
+        }
+        if (mnList.HasUniqueProperty(pl.pubKeyOperator)) {
+            return state.DoS(10, false, REJECT_DUPLICATE, "bad-protx-dup-operator-key");
+        }
+    }
+
+    return true;
+}
+
+// Provider Update Service Payload
+static bool CheckProUpServTx(const CTransaction& tx, const CBlockIndex* pindexPrev, CValidationState& state)
+{
+    assert(tx.nType == CTransaction::TxType::PROUPSERV);
+
+    ProUpServPL pl;
+    if (!GetTxPayload(tx, pl)) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-payload");
+    }
+
+    if (pl.nVersion == 0 || pl.nVersion > ProUpServPL::CURRENT_VERSION) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-version");
+    }
+
+    if (!CheckService(pl.addr, state)) {
+        // pass the state returned by the function above
+        return false;
+    }
+
+    if (!CheckInputsHash(tx, pl, state)) {
+        // pass the state returned by the function above
+        return false;
+    }
+
+    if (pindexPrev) {
+        auto mnList = deterministicMNManager->GetListForBlock(pindexPrev);
+        auto mn = mnList.GetMN(pl.proTxHash);
+        if (!mn) {
+            return state.DoS(100, false, REJECT_INVALID, "bad-protx-hash");
+        }
+
+        // don't allow updating to addresses already used by other MNs
+        if (mnList.HasUniqueProperty(pl.addr) && mnList.GetUniquePropertyMN(pl.addr)->proTxHash != pl.proTxHash) {
+            return state.DoS(10, false, REJECT_DUPLICATE, "bad-protx-dup-addr");
+        }
+
+        if (!pl.scriptOperatorPayout.empty()) {
+            if (mn->nOperatorReward == 0) {
+                // don't allow to set operator reward payee in case no operatorReward was set
+                return state.DoS(10, false, REJECT_INVALID, "bad-protx-operator-payee");
+            }
+            // we may support other kinds of scripts later, but restrict it for now
+            if (!pl.scriptOperatorPayout.IsPayToPublicKeyHash()) {
+                return state.DoS(10, false, REJECT_INVALID, "bad-protx-operator-payee");
+            }
+        }
+
+        // we can only check the signature if pindexPrev != nullptr and the MN is known
+        if (!CheckHashSig(pl, mn->pdmnState->pubKeyOperator.Get(), state)) {
+            // pass the state returned by the function above
+            return false;
+        }
+    }
+
+    return true;
+}
+
+// Provider Update Registrar Payload
+static bool CheckProUpRegTx(const CTransaction& tx, const CBlockIndex* pindexPrev, const CCoinsViewCache* view, CValidationState& state)
+{
+    assert(tx.nType == CTransaction::TxType::PROUPREG);
+
+    ProUpRegPL pl;
+    if (!GetTxPayload(tx, pl)) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-payload");
+    }
+
+    if (pl.nVersion == 0 || pl.nVersion > ProUpRegPL::CURRENT_VERSION) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-version");
+    }
+    if (pl.nMode != 0) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-mode");
+    }
+
+    if (!pl.pubKeyOperator.IsValid()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-operator-key-invalid");
+    }
+    if (pl.keyIDVoting.IsNull()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-voting-key-null");
+    }
+    // !TODO: enable other scripts
+    if (!pl.scriptPayout.IsPayToPublicKeyHash()) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-payee");
+    }
+
+    CTxDestination payoutDest;
+    if (!ExtractDestination(pl.scriptPayout, payoutDest)) {
+        // should not happen as we checked script types before
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-payee-dest");
+    }
+
+    // don't allow reuse of payee key for other keys
+    if (payoutDest == CTxDestination(pl.keyIDVoting)) {
+        return state.DoS(10, false, REJECT_INVALID, "bad-protx-payee-reuse");
+    }
+
+    if (!CheckInputsHash(tx, pl, state)) {
+        return false;
+    }
+
+    if (pindexPrev) {
+        assert(view != nullptr);
+
+        // ProUpReg txes are disabled when the legacy system is still active
+        // !TODO: remove after complete transition to DMN
+        if (!deterministicMNManager->LegacyMNObsolete(pindexPrev->nHeight + 1)) {
+            return state.DoS(10, false, REJECT_INVALID, "spork-21-inactive");
+        }
+
+        auto mnList = deterministicMNManager->GetListForBlock(pindexPrev);
+        auto dmn = mnList.GetMN(pl.proTxHash);
+        if (!dmn) {
+            return state.DoS(100, false, REJECT_INVALID, "bad-protx-hash");
+        }
+
+        // don't allow reuse of payee key for owner key
+        if (payoutDest == CTxDestination(dmn->pdmnState->keyIDOwner)) {
+            return state.DoS(10, false, REJECT_INVALID, "bad-protx-payee-reuse");
+        }
+
+        Coin coin;
+        if (!view->GetUTXOCoin(dmn->collateralOutpoint, coin)) {
+            // this should never happen (there would be no dmn otherwise)
+            return state.DoS(100, false, REJECT_INVALID, "bad-protx-collateral");
+        }
+
+        // don't allow reuse of collateral key for other keys (don't allow people to put the payee key onto an online server)
+        CTxDestination collateralTxDest;
+        if (!ExtractDestination(coin.out.scriptPubKey, collateralTxDest)) {
+            return state.DoS(100, false, REJECT_INVALID, "bad-protx-collateral-dest");
+        }
+        if (collateralTxDest == CTxDestination(dmn->pdmnState->keyIDOwner) ||
+                collateralTxDest == CTxDestination(pl.keyIDVoting)) {
+            return state.DoS(10, false, REJECT_INVALID, "bad-protx-collateral-reuse");
+        }
+
+        if (mnList.HasUniqueProperty(pl.pubKeyOperator)) {
+            auto otherDmn = mnList.GetUniquePropertyMN(pl.pubKeyOperator);
+            if (pl.proTxHash != otherDmn->proTxHash) {
+                return state.DoS(10, false, REJECT_DUPLICATE, "bad-protx-dup-key");
+            }
+        }
+
+        if (!CheckHashSig(pl, dmn->pdmnState->keyIDOwner, state)) {
+            // pass the state returned by the function above
+            return false;
+        }
+
+    }
+
+    return true;
+}
+
+// Provider Update Revoke Payload
+static bool CheckProUpRevTx(const CTransaction& tx, const CBlockIndex* pindexPrev, CValidationState& state)
+{
+    assert(tx.nType == CTransaction::TxType::PROUPREV);
+
+    ProUpRevPL pl;
+    if (!GetTxPayload(tx, pl)) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-payload");
+    }
+
+    if (pl.nVersion == 0 || pl.nVersion > ProUpRevPL::CURRENT_VERSION) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-version");
+    }
+
+    // pl.nReason < ProUpRevPL::REASON_NOT_SPECIFIED is always `false` since
+    // pl.nReason is unsigned and ProUpRevPL::REASON_NOT_SPECIFIED == 0
+    if (pl.nReason > ProUpRevPL::REASON_LAST) {
+        return state.DoS(100, false, REJECT_INVALID, "bad-protx-reason");
+    }
+
+    if (!CheckInputsHash(tx, pl, state)) {
+        // pass the state returned by the function above
+        return false;
+    }
+
+    if (pindexPrev) {
+        auto mnList = deterministicMNManager->GetListForBlock(pindexPrev);
+        auto dmn = mnList.GetMN(pl.proTxHash);
+        if (!dmn)
+            return state.DoS(100, false, REJECT_INVALID, "bad-protx-hash");
+
+        if (!CheckHashSig(pl, dmn->pdmnState->pubKeyOperator.Get(), state)) {
+            // pass the state returned by the function above
+            return false;
+        }
+    }
+
+    return true;
+}
 
 // Basic non-contextual checks for all tx types
 static bool CheckSpecialTxBasic(const CTransaction& tx, CValidationState& state)
