+++ test/lint/lint-circular-dependencies.sh
@@ -19,9 +19,7 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "chainparamsbase -> util/system -> chainparamsbase"
     "consensus/params -> consensus/upgrades -> consensus/params"
     "crypter -> wallet/wallet -> crypter"
-    "evo/deterministicmns -> evo/providertx -> evo/deterministicmns"
     "evo/deterministicmns -> masternodeman -> evo/deterministicmns"
-    "evo/providertx -> evo/specialtx -> evo/providertx"
     "kernel -> validation -> kernel"
     "masternode -> masternode-sync -> masternode"
     "masternode -> masternodeman -> masternode"
@@ -56,7 +54,6 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "chainparamsbase -> util/system -> logging -> chainparamsbase"
     "coins -> policy/fees -> txmempool -> coins"
     "consensus/zerocoin_verify -> spork -> validation -> consensus/zerocoin_verify"
-    "evo/deterministicmns -> evo/providertx -> evo/specialtx -> evo/deterministicmns"
     "evo/deterministicmns -> masternode -> masternode-sync -> evo/deterministicmns"
     "evo/deterministicmns -> masternodeman -> net_processing -> evo/deterministicmns"
     "evo/deterministicmns -> masternode -> wallet/wallet -> evo/deterministicmns"
@@ -77,9 +74,10 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "chain -> legacy/stakemodifier -> validation -> zpivchain -> chain"
     "coins -> policy/fees -> policy/policy -> validation -> coins"
     "coins -> policy/fees -> txmempool -> evo/deterministicmns -> coins"
-    "coins -> policy/fees -> txmempool -> evo/providertx -> coins"
+    "coins -> policy/fees -> txmempool -> evo/specialtx -> coins"
     "consensus/tx_verify -> consensus/zerocoin_verify -> spork -> validation -> consensus/tx_verify"
     "evo/deterministicmns -> masternodeman -> net_processing -> validationinterface -> evo/deterministicmns"
+    "evo/deterministicmns -> masternodeman -> validation -> evo/specialtx -> evo/deterministicmns"
     "coins -> policy/fees -> policy/policy -> validation -> txdb -> coins"
     "chain -> legacy/stakemodifier -> stakeinput -> wallet/wallet -> spork -> net_processing -> chain"
 )
